# CCT: Improve performance of Valid Math Expressions checker