As the title says. Test build at https://yichizhng.github.io/bitburner-src/

Script for testing:
```js
function helper(res, path, num, target, pos, evaluated, multed) {
    if (pos === num.length) {
        if (target === evaluated) {
            res.push(path);
        }
        return;
    }

    for (let i = pos; i < num.length; ++i) {
        if (i != pos && num[pos] == "0") {
            break;
        }
        const cur = parseInt(num.substring(pos, i + 1));

        if (pos === 0) {
            helper(res, path + cur, num, target, i + 1, cur, cur);
        } else {
            helper(res, path + "+" + cur, num, target, i + 1, evaluated + cur, cur);
            helper(res, path + "-" + cur, num, target, i + 1, evaluated - cur, -cur);
            helper(res, path + "*" + cur, num, target, i + 1, evaluated - multed + multed * cur, multed * cur);
        }
    }
}

/** @param {NS} ns */
export async function main(ns) {
    ns.clearLog();
    ns.disableLog('ALL');
    for (let i = 0; i < 1000; ++i) {
        ns.codingcontract.createDummyContract(
            "Find All Valid Math Expressions"
        );
    }
    let files = ns.ls('home', '.cct');
    let start = performance.now();
    let answers = new Map();
    for (let file of files) {
        const result = [];
        let [n, t] = ns.codingcontract.getData(file);
        helper(result, "", n, t, 0, 0, 0);
        answers.set(file, result);
    }
    ns.print(`generated answers in ${performance.now() - start}ms`);
    start = performance.now();
    for (let file of files) {
        ns.codingcontract.attempt(answers.get(file), file);
    }
    ns.print(`checked answers in ${performance.now() - start}ms`);
}
```

closes #1095