@d0sboots 2024-05-18
Does `Find All Valid Math Expressions` have a test? What you wrote there seems like it would be straightforward to adapt (we might not want 1000 iterations, depending on speed), and if we don't have a test for this we really ought to.

If you really don't want to (or it turns into a big pain) I'll take the improvement as-is.