@yichizhng 2024-05-19
That script just uses the same code that the game uses to check the solution, so I don't think that it has any value as a test for correctness. A test for performance has too much variance due to the randomness in input size.

The point of the test script is that the difference in the two times printed is the amount of time spent comparing the results (plus whatever other overhead). Some sample results on Chrome (Firefox has better performance on Array and worse on Set, but it's still an improvement)

Before:
```
generated answers in 12703.600000023842ms
checked answers in 29619.900000035763ms
```

After:
```
generated answers in 11085.5ms
checked answers in 11453.300000011921ms
```